requirements.txt
pandas
seaborn
matplotlib
ydata-profiling
scikit-learn
streamlit
joblibscikit-learn
streamlit


